requirements.txt
flask
Pillow
torchvision
numpy
opencv-python-headless
panda
fiftyone